requirements.txt
altair==5.5.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.11
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.1.0
folium==0.19.5
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.31.0
nest_asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.1.0
pip==25.0
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.2
python-dateutil==2.9.0.post0
pytz==2025.1
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
stack_data==0.6.3
streamlit==1.43.2
streamlit_folium==0.24.0
tenacity==9.0.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
xgboost==3.0.0
xyzservices==2025.1.0
zipp==3.21.0
